organize gpu modules and document exported api
