+++ app/packages/multimodal/src/adapters/mcap/decoders/foxglove/protobuf/payloads.ts
@@ -10,7 +10,7 @@ export const FOXGLOVE_COMPRESSED_IMAGE_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.CompressedImage",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identity for foxglove.CompressedVideo messages.
@@ -19,7 +19,7 @@ export const FOXGLOVE_COMPRESSED_VIDEO_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.CompressedVideo",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identity for foxglove.RawImage messages.
@@ -28,7 +28,7 @@ export const FOXGLOVE_RAW_IMAGE_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.RawImage",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identity for foxglove.PointCloud messages.
@@ -37,7 +37,7 @@ export const FOXGLOVE_POINT_CLOUD_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.PointCloud",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identity for foxglove.ImageAnnotations messages.
@@ -46,7 +46,7 @@ export const FOXGLOVE_IMAGE_ANNOTATIONS_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.ImageAnnotations",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identity for foxglove.LaserScan messages.
@@ -55,7 +55,7 @@ export const FOXGLOVE_LASER_SCAN_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.LaserScan",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identity for foxglove.SceneUpdate messages.
@@ -64,7 +64,7 @@ export const FOXGLOVE_SCENE_UPDATE_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.SceneUpdate",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identity for foxglove.Grid messages.
@@ -73,7 +73,7 @@ export const FOXGLOVE_GRID_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.Grid",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identity for foxglove.CameraCalibration messages.
@@ -82,7 +82,7 @@ export const FOXGLOVE_CAMERA_CALIBRATION_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.CameraCalibration",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identity for foxglove.PoseInFrame messages.
@@ -91,7 +91,7 @@ export const FOXGLOVE_POSE_IN_FRAME_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.PoseInFrame",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identity for foxglove.LocationFix messages.
@@ -100,7 +100,7 @@ export const FOXGLOVE_LOCATION_FIX_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.LocationFix",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identity for foxglove.Log messages.
@@ -109,7 +109,7 @@ export const FOXGLOVE_LOG_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.Log",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 /**
  * Payload identities for Foxglove frame transform protobuf messages.
@@ -118,10 +118,10 @@ export const FOXGLOVE_FRAME_TRANSFORM_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.FrameTransform",
   schemaEncoding: "protobuf",
-};
+} as const;
 
 export const FOXGLOVE_FRAME_TRANSFORMS_PAYLOAD: PayloadDescriptor = {
   encoding: "protobuf",
   schema: "foxglove.FrameTransforms",
   schemaEncoding: "protobuf",
-};
+} as const;
