+++ app/packages/multimodal/src/visualization/panels/gpu/gpu-pick-readback-pool.test.ts
