+++ app/packages/multimodal/src/visualization/panels/gpu/gpu-point-cloud-projection-picker.test.tsx
