+++ app/packages/multimodal/src/visualization/panels/gpu/gpu-point-cloud-projection.ts
@@ -1,9 +1,10 @@
 import * as THREE from "three";
 
-import type { ImageViewTransform } from "./base-2d-scene";
+import type { ImageViewTransform } from "../base-2d-scene";
 
 const QUATERNION_EPSILON = 1e-9;
 
+/** Camera intrinsics and image dimensions consumed by GPU projection math. */
 export interface GpuProjectionCalibration {
   readonly K?: readonly number[] | null;
   readonly P?: readonly number[] | null;
@@ -11,6 +12,7 @@ export interface GpuProjectionCalibration {
   readonly width: number;
 }
 
+/** Sensor-to-camera rotation represented as an xyzw quaternion. */
 export interface GpuProjectionRotation {
   readonly w: number;
   readonly x: number;
@@ -18,6 +20,7 @@ export interface GpuProjectionRotation {
   readonly z: number;
 }
 
+/** Sensor-to-camera translation in metres. */
 export interface GpuProjectionTranslation {
   readonly x: number;
   readonly y: number;
@@ -118,6 +121,7 @@ export function gpuProjectionImagePlaneSize({
     : { height: safeContainerHeight, width: safeContainerHeight * imageAspect };
 }
 
+/** Projected image bounds in normalized viewport coordinates. */
 export interface GpuProjectionViewportRect {
   readonly bottom: number;
   readonly left: number;
