+++ app/packages/multimodal/src/visualization/panels/gpu/webgpu-canvas.test.tsx
