+++ app/packages/multimodal/src/visualization/panels/gpu/webgpu-device-registry.test.ts
