+++ app/packages/multimodal/src/visualization/panels/gpu/webgpu-view-stage.test.tsx
