+++ app/packages/multimodal/src/visualization/panels/point-cloud/gpu/gpu-point-cloud-3d-picker.test.ts
