+++ app/packages/multimodal/src/visualization/panels/point-cloud/gpu/gpu-point-cloud-color.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { buildPointCloudRenderPayload } from "../../../decoders";
+import { buildPointCloudRenderPayload } from "../../../../decoders";
 import {
   gpuPointCloudColorAtSample,
   NEUTRAL_GPU_POINT_COLOR,
