+++ app/packages/multimodal/src/visualization/panels/point-cloud/gpu/gpu-point-cloud-sampling.ts
